# api/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"test","description":"","version":"1.0.0"},"tags":[],"paths":{"/tags":{"get":{"summary":"全部标签","deprecated":false,"description":"","operationId":"GetAllTags","tags":[],"parameters":[],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{},"APIPOST_ORDERS":["code","msg","data"],"APIPOST_REFS":{}}}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[]}},"/categories":{"get":{"summary":"全部分类","deprecated":false,"description":"","operationId":"GetAllCategorys","tags":[],"parameters":[],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{},"APIPOST_ORDERS":["code","msg","data"],"APIPOST_REFS":{}}}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[]},"post":{"summary":"创建分类","deprecated":false,"description":"","operationId":"CreateCategory","tags":[],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":"名称"},"desc":{"type":"string"}},"required":["name","desc"]},"example":{"name":"探索日志","desc":"这是一段简介"}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]}},"/token":{"get":{"summary":"获取管理员token","deprecated":false,"description":"","operationId":"GetAdminToken","tags":[],"parameters":[{"name":"key","in":"query","description":"","required":false,"example":"asfadsfasfe","schema":{"type":"string"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[]}},"/post/{uri}":{"get":{"summary":"帖子","deprecated":false,"description":"","operationId":"GetPostByUri","tags":[],"parameters":[{"name":"uri","in":"path","description":"","required":true,"example":"test-uri1","schema":{"type":"string"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{},"APIPOST_ORDERS":["code","msg","data"],"APIPOST_REFS":{}}}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[]}},"/posts":{"get":{"summary":"帖子列表","deprecated":false,"description":"","operationId":"GetPostList","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"2","schema":{"type":"integer"}},{"name":"tag","in":"query","description":"","required":false,"example":"a","schema":{"type":"string"}},{"name":"category","in":"query","description":"","required":false,"example":"95","schema":{"type":"integer","format":"uint32"}},{"name":"limit","in":"query","description":"","required":false,"example":"13","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{},"APIPOST_ORDERS":["code","msg","data"],"APIPOST_REFS":{}}}}},"404":{"description":"失败","content":{"application/json":{"schema":{"type":"string"}}}}},"security":[]},"post":{"summary":"创建Post","deprecated":false,"description":"","operationId":"CreatePost","tags":[],"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"uri":{"description":"ID 编号","example":"test1","type":"string"},"content":{"format":"binary","type":"string","example":"file://D:\\MarkDown\\blog\\test.md"}},"required":["uri","content"]}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]}},"/posts/recent":{"get":{"summary":"最近帖子","deprecated":false,"description":"","operationId":"GetRecentPostList","tags":[],"parameters":[],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{},"required":["01HXCAZFYE7XPG6HMRQYM1PGAE"]}}}}},"security":[]}},"/post/{id}":{"put":{"summary":"修改Post","deprecated":false,"description":"","operationId":"ModifyPosts","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"example":"2000","schema":{"type":"integer","format":"uint32"}}],"requestBody":{"content":{"application/json":{"schema":{"properties":{"title":{"type":"string"},"category":{"type":"integer","description":"分组","format":"uint32"},"tags":{"type":"array","items":{"type":"string"}},"visible":{"type":"boolean"},"desc":{"type":"string"}},"type":"object"},"example":"{\r\n    \"title\": \"测试文档1\",\r\n    \"category\": 95,\r\n    \"tags\": [\r\n        \"markdown\",\r\n        \"test\",\r\n        \"tag1\",\r\n        \"tag2\",\r\n    ],\r\n    \"visible\": true,\r\n    \"desc\": \"这段 Markdown 文档是一个全面测试 Markdown 语法的示例文档。\\n\\n它包含了常见的 Markdown 元素，如段落、列表、链接、引用、代码块、表格、图片等。每个元素都有相应的示例，方便查看 Markdown 的渲染效果。\"\r\n}"}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]},"delete":{"summary":"删除","deprecated":false,"description":"","operationId":"DeletePost","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1980","schema":{"type":"integer","format":"uint32"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]},"post":{"summary":"设置PostContent","deprecated":false,"description":"","operationId":"SetPostContent","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"example":"1990","schema":{"type":"integer","format":"uint32"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"content":{"format":"binary","type":"string","example":"file://D:\\MarkDown\\blog\\test.md"}},"required":["content"]}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]}},"/category/{id}":{"patch":{"summary":"修改简介","deprecated":false,"description":"","operationId":"ModifyCategoryDesc","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"schema":{"type":"integer","format":"uint32"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"desc":{"type":"string"}},"required":["desc"]}}}},"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]},"delete":{"summary":"删除分类","deprecated":false,"description":"","operationId":"DeleteCategory","tags":[],"parameters":[{"name":"id","in":"path","description":"","required":true,"example":"94","schema":{"type":"integer","format":"uint32"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]}},"/admin/posts":{"get":{"summary":"Admin Get PostList","deprecated":false,"description":"","operationId":"AdminGetPostList","tags":[],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"integer"}}],"responses":{"200":{"description":"成功","content":{"application/json":{"schema":{"type":"object","properties":{}}}}}},"security":[{"bearer":[]}]}}},"components":{"schemas":{},"securitySchemes":{"bearer":{"type":"http","scheme":"bearer"}}},"servers":[]}